+++ src/components/build_intermediate_dependencies.ts
@@ -1,7 +1,9 @@
 import chalk from "chalk";
 import * as fs from "fs";
 import * as path from "path";
+import { BuildStatusConfig } from "../config/buildStatus/BuildStatusConfig";
 import { RearguardConfig } from "../config/rearguard/RearguardConfig";
+import { NON_VERSIONABLE_CONFIG_FILE_NAME } from "../const";
 import { IEnvConfig } from "../interfaces/config/IEnvConfig";
 import { IRearguardConfig } from "../interfaces/config/IRearguardConfig";
 import { build } from "./mono_repository/build";
@@ -25,32 +27,39 @@ export async function build_intermediate_dependencies(
 
     if (fs.existsSync(i_global_dep)) {
       const i_rearguardConfig = new RearguardConfig(envConfig, path.resolve(i_global_dep, "package.json"));
+      const i_buildStatusConfig = new BuildStatusConfig(path.resolve(i_global_dep, NON_VERSIONABLE_CONFIG_FILE_NAME));
       const i_deps = i_rearguardConfig.sync_project_deps;
-      const i_build_time = i_rearguardConfig.last_build_time;
+      const i_has_last_build_time = i_buildStatusConfig.has_last_build_time;
 
-      for (let k = i_deps.length - 1; k >= 0; k--) {
-        const k_dep_name = i_deps[k];
-        const k_global_dep = envConfig.resolveGlobalModule(k_dep_name);
+      if (i_has_last_build_time) {
+        const i_build_time = i_buildStatusConfig.last_build_time;
 
-        if (fs.existsSync(k_global_dep)) {
-          const k_rearguardConfig = new RearguardConfig(
-            envConfig,
-            path.resolve(envConfig.resolveGlobalModule(k_dep_name), "package.json"),
-          );
-          const k_build_time = k_rearguardConfig.last_build_time;
-          const k_has_last_build_time = k_rearguardConfig.has_last_build_time;
+        for (let k = i_deps.length - 1; k >= 0; k--) {
+          const k_dep_name = i_deps[k];
+          const k_global_dep = envConfig.resolveGlobalModule(k_dep_name);
 
-          if (k_has_last_build_time) {
-            if (i_build_time.isSameOrBefore(k_build_time)) {
+          if (fs.existsSync(k_global_dep)) {
+            const k_buildStatusConfig = new BuildStatusConfig(
+              path.resolve(k_global_dep, NON_VERSIONABLE_CONFIG_FILE_NAME),
+            );
+            const k_has_last_build_time = k_buildStatusConfig.has_last_build_time;
+
+            if (k_has_last_build_time) {
+              const k_build_time = k_buildStatusConfig.last_build_time;
+
+              if (i_build_time.isSameOrBefore(k_build_time)) {
+                CWD_list.push(i_global_dep);
+                break;
+              }
+            } else if (!CWD_list.includes(i_global_dep)) {
               CWD_list.push(i_global_dep);
-              break;
             }
-          } else if (!CWD_list.includes(i_global_dep)) {
-            CWD_list.push(i_global_dep);
+          } else {
+            console.log(chalk.bold.grey(`[ GLOBAL_MODULE: ${k_dep_name} ][ NOT_FOUND ]`));
           }
-        } else {
-          console.log(chalk.bold.grey(`[ GLOBAL_MODULE: ${k_dep_name} ][ NOT_FOUND ]`));
         }
+      } else if (!CWD_list.includes(i_global_dep)) {
+        CWD_list.push(i_global_dep);
       }
     } else if (fs.existsSync(i_local_dep)) {
       console.log(chalk.grey(`[ MODULE: ${i_dep_name} ][ LOCAL_INSTALED ]`));
@@ -60,8 +69,8 @@ export async function build_intermediate_dependencies(
   }
 
   if (CWD_list.length > 0) {
-    console.log(chalk.bold.white(`[ TARGET_LIST ]`));
-    console.log(chalk.bold.white(JSON.stringify(CWD_list, null, 2)));
+    console.log(chalk.bold.yellowBright(`[ TARGET_LIST ]`));
+    console.log(chalk.bold.yellowBright(JSON.stringify(CWD_list, null, 2)));
   } else {
     console.log(chalk.grey(`[ HAVE_NOT_TARGETS ]`));
   }