+++ src/webpack/webpack.config.dll.ts
@@ -1,9 +1,10 @@
 import * as path from "path";
+import { buildStatusConfig } from "../config/buildStatus";
 import { DLL_BUNDLE_DIR_NAME } from "../const";
 import { dll_entry_name, dll_output_path, get_context } from "../helpers";
 import { IEnvConfig } from "../interfaces/config/IEnvConfig";
 import { IRearguardConfig } from "../interfaces/config/IRearguardConfig";
-import { analyze, assetsPlugin, clean, DllPlugin } from "./components/js.plugins";
+import { analyze, assetsPlugin, clean, DllPlugin, HashWebpackPlugin } from "./components/js.plugins";
 import tsLoader from "./components/ts.loaders";
 import { general_WP_config } from "./webpack.config.common";
 
@@ -33,6 +34,7 @@ export function dll_WP_config(envConfig: IEnvConfig, rearguardConfig: IRearguard
       ...assetsPlugin(envConfig, DLL_BUNDLE_DIR_NAME),
       ...analyze(envConfig),
       ...clean(envConfig, [dll_output_path()]),
+      new HashWebpackPlugin(buildStatusConfig, envConfig),
     ],
     {},
   );